fix warnings

+++ source/input/I2CKeyboardInputDriver.cpp
@@ -81,7 +81,7 @@ TDeckKeyboardInputDriver::TDeckKeyboardInputDriver(uint8_t address)
 void TDeckKeyboardInputDriver::readKeyboard(uint8_t address, lv_indev_t *indev, lv_indev_data_t *data)
 {
     char keyValue = 0;
-    uint8_t bytes = Wire.requestFrom(address, 1);
+    uint8_t bytes = Wire.requestFrom(address, (uint8_t)1);
     if (Wire.available() > 0 && bytes > 0) {
         keyValue = Wire.read();
         // ignore empty reads and keycode 224(E0, shift-0 on T-Deck) which causes internal issues
@@ -429,7 +429,7 @@ void BBQ10KeyboardInputDriver::init(void)
 void BBQ10KeyboardInputDriver::readKeyboard(uint8_t address, lv_indev_t *indev, lv_indev_data_t *data)
 {
     char keyValue = 0;
-    uint8_t bytes = Wire.requestFrom(address, 1);
+    uint8_t bytes = Wire.requestFrom(address, (uint8_t)1);
     if (Wire.available() > 0 && bytes > 0) {
         keyValue = Wire.read();
         // ignore empty reads and keycode 224(E0, shift-0 on T-Deck) which causes internal issues
@@ -461,7 +461,7 @@ CardKBInputDriver::CardKBInputDriver(uint8_t address)
 void CardKBInputDriver::readKeyboard(uint8_t address, lv_indev_t *indev, lv_indev_data_t *data)
 {
     char keyValue = 0;
-    Wire.requestFrom(address, 1);
+    Wire.requestFrom(address, (uint8_t)1);
     if (Wire.available() > 0) {
         keyValue = Wire.read();
         // ignore empty reads and keycode 224 which causes internal issues
